pyproject.toml
[tool.poetry]
name = "gcp-cloud-run-template-api"
version = "0.1.0"
description = ""
authors = ["your.name <your.name@exmaple.co.uk>"]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
fastapi = "==0.103.1"
gunicorn = "==21.2.0"
uvicorn = "==0.23.2"
pendulum = "^2.1.2"
google-cloud-pubsub = "==2.18.3"
google-cloud-storage = "==2.10.0"
google-cloud-secret-manager = "==2.16.1"
gcp-storage-emulator = "==2022.6.11"
python-json-logger = "^2.0.7"
google-cloud-datastore = "==2.19.0"
pydantic = {version = "==2.3.0", extras = ["email"]}
pydantic-settings = "==2.0.3"
PyYAML = "==6.0.1"
datamodel-code-generator = "==0.25.5"

[tool.poetry.dev-dependencies]
pytest = "==7.4.1"
httpx="*"
requests_mock = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
